requirements.txt
pillow
opencv
numpy
pandas
keras
tkinter